fix: move dependencies
