+++ app/dependencies/code_genration_service.py
@@ -1,0 +1,10 @@
+from typing import Annotated
+
+from fastapi import Depends
+
+from app.services.code_generation import CodeGenerationService
+
+CodeGenerationServiceDep = Annotated[
+    CodeGenerationService,
+    Depends(CodeGenerationService),
+]
