+++ app/services/code_generation.py
@@ -1,47 +1,25 @@
-from enum import Enum
 from uuid import UUID
 
+from app.dependencies.utils import (
+    GeneratorFactoryDep,
+    UMLGraphPreloaderServiceDep,
+)
 from app.services.uml_graph_preloader import UMLGraphPreloaderService
-from app.utils.code_generator import JavaCodeGenerator, PythonCodeGenerator
+from app.utils.generator_factory import GeneratorFactory, TargetLanguage
 from app.utils.uml_analyzer import UMLGraphAnalyzer
 
 
-class TargetLanguage(str, Enum):
-    JAVA = 'java'
-    PYTHON = 'python'
-
-
-class GeneratorFactory:
-    def __init__(
-        self,
-        java_generator: JavaCodeGenerator,
-        python_generator: PythonCodeGenerator,
-    ):
-        self._java_generator = java_generator
-        self._python_generator = python_generator
-
-    def get(self, language: TargetLanguage):
-        mapping = {
-            TargetLanguage.JAVA: self._java_generator,
-            TargetLanguage.PYTHON: self._python_generator,
-        }
-
-        generator = mapping.get(language)
-
-        if generator is None:
-            raise ValueError(f'Unsupported language: {language}')
-
-        return generator
-
-
 class CodeGenerationService:
+    __graph_loader: UMLGraphPreloaderService
+    __generator_factory: GeneratorFactory
+
     def __init__(
         self,
-        graph_loader: UMLGraphPreloaderService,
-        generator_factory: GeneratorFactory,
+        graph_loader: UMLGraphPreloaderServiceDep,
+        generator_factory: GeneratorFactoryDep,
     ):
-        self._graph_loader = graph_loader
-        self._generator_factory = generator_factory
+        self.__graph_loader = graph_loader
+        self.__generator_factory = generator_factory
 
     async def generate(
         self,
@@ -49,11 +27,11 @@ class CodeGenerationService:
         language: TargetLanguage,
     ) -> dict[str, str]:
 
-        graph = await self._graph_loader.load_window_graph(window_id)
+        graph = await self.__graph_loader.load_window_graph(window_id)
 
         analyzer = UMLGraphAnalyzer(graph)
 
-        generator = self._generator_factory.get(language)
+        generator = self.__generator_factory.get(language)
 
         files: dict[str, str] = {}
